functions/appointment_checker/requirements.txt
functions-framework==3.5.0
google-cloud-bigquery==3.17.2